Refactor security validation: Remove redundant functions and introduce new modular files
- Removed legacy validation functions from `validator.go`.
- Added `types.go` and `validator.go` in `internal/runner/security` for
  modular security validation.
- Introduced new error definitions, logging options, and configuration
  structures.
- Enhanced regex-based validation for commands and environment
  variables.
